flows: move markers to separate files, cleanup formatting
