+++ passbook/stages/captcha/tests.py
@@ -5,8 +5,9 @@ from django.test import Client, TestCase
 from django.utils.encoding import force_text
 
 from passbook.core.models import User
+from passbook.flows.markers import StageMarker
 from passbook.flows.models import Flow, FlowDesignation, FlowStageBinding
-from passbook.flows.planner import FlowPlan, StageMarker
+from passbook.flows.planner import FlowPlan
 from passbook.flows.views import SESSION_KEY_PLAN
 from passbook.stages.captcha.models import CaptchaStage
