+++ passbook/stages/invitation/tests.py
@@ -7,8 +7,9 @@ from django.utils.encoding import force_text
 from guardian.shortcuts import get_anonymous_user
 
 from passbook.core.models import User
+from passbook.flows.markers import StageMarker
 from passbook.flows.models import Flow, FlowDesignation, FlowStageBinding
-from passbook.flows.planner import PLAN_CONTEXT_PENDING_USER, FlowPlan, StageMarker
+from passbook.flows.planner import PLAN_CONTEXT_PENDING_USER, FlowPlan
 from passbook.flows.views import SESSION_KEY_PLAN
 from passbook.stages.invitation.forms import InvitationStageForm
 from passbook.stages.invitation.models import Invitation, InvitationStage
