+++ passbook/stages/user_write/tests.py
@@ -7,8 +7,9 @@ from django.test import Client, TestCase
 from django.utils.encoding import force_text
 
 from passbook.core.models import User
+from passbook.flows.markers import StageMarker
 from passbook.flows.models import Flow, FlowDesignation, FlowStageBinding
-from passbook.flows.planner import PLAN_CONTEXT_PENDING_USER, FlowPlan, StageMarker
+from passbook.flows.planner import PLAN_CONTEXT_PENDING_USER, FlowPlan
 from passbook.flows.views import SESSION_KEY_PLAN
 from passbook.stages.prompt.stage import PLAN_CONTEXT_PROMPT
 from passbook.stages.user_write.forms import UserWriteStageForm
